Extract AO for ensuring current lineup
