+++ app/jobs/ensure_current_lineup_job.rb
@@ -1,0 +1,5 @@
+class EnsureCurrentLineupJob < ApplicationJob
+  def perform
+    Lineup::Builder.ensure_current
+  end
+end
