+++ app/models/lineup/builder.rb
@@ -1,5 +1,5 @@
-class CreateLineupJob < ApplicationJob
-  def perform
+class Lineup::Builder
+  def self.ensure_current
     Lineup.create_next while Lineup.current.nil?
     Lineup.create_next
   end
